# gen/typescript/api.schemas.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Category API
 * API for querying categories
 * OpenAPI spec version: 1.1.0
 */
/**
 * How an attribute is used in a category
 */
export type AttributeRole = typeof AttributeRole[keyof typeof AttributeRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AttributeRole = {
  variant: 'variant',
  specification: 'specification',
} as const;

/**
 * Type of the attribute
 */
export type AttributeType = typeof AttributeType[keyof typeof AttributeType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AttributeType = {
  single: 'single',
  multiple: 'multiple',
  range: 'range',
  boolean: 'boolean',
  text: 'text',
} as const;

/**
 * Attribute option for single/multiple type attributes
 */
export interface AttributeOption {
  /** Display name of the option */
  name: string;
  /** URL-friendly identifier */
  slug: string;
  /** Color hex code (for color attributes) */
  colorCode?: string;
  /** Sort order for display */
  sortOrder: number;
}

export interface CategoryAttribute {
  /** Unique attribute identifier (UUID) */
  attributeId: string;
  /** Display name of the attribute (denormalized from catalog-service) */
  name: string;
  /** URL-friendly identifier (denormalized from catalog-service) */
  slug: string;
  type: AttributeType;
  /** Unit of measurement for range attributes (denormalized from catalog-service) */
  unit?: string;
  /** Available options for single/multiple type attributes (denormalized from catalog-service) */
  options?: AttributeOption[];
  role: AttributeRole;
  /** Sort order for display */
  sortOrder: number;
  /** Whether the attribute is filterable for this category */
  filterable: boolean;
  /** Whether the attribute is searchable for this category */
  searchable: boolean;
}

export interface CategoryResponse {
  id: string;
  name: string;
  attributes: CategoryAttribute[];
}

export type ProblemErrorsItem = {
  field?: string;
  message?: string;
};

/**
 * RFC7807 Problem Details
 */
export interface Problem {
  type: string;
  title: string;
  status: number;
  detail?: string;
  instance?: string;
  traceId?: string;
  errors?: ProblemErrorsItem[];
}